+++ packages/apollo-cache-inmemory/src/index.ts
@@ -1,0 +1,7 @@
+import { InMemoryCache, defaultDataIdFromObject } from './inMemoryCache';
+export * from './readFromStore';
+export * from './writeToStore';
+export * from './fragmentMatcher';
+export * from './types';
+export { defaultDataIdFromObject };
+export default InMemoryCache;