+++ packages/apollo-cache-inmemory/tests/tests.ts
@@ -1,0 +1,4 @@
+import './cache';
+import './diffAgainstStore';
+import './readFromStore';
+import './writeToStore';