+++ packages/apollo-client/src/util/errorHandling.ts
@@ -1,9 +1,0 @@
-export function tryFunctionOrLogError(f: Function) {
-  try {
-    return f();
-  } catch (e) {
-    if (console.error) {
-      console.error(e);
-    }
-  }
-}