+++ packages/apollo-client/CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change log
 
 ### vNEXT
+- Replace core utils with apollo-utilities
 - Cleanup InMemoryCache to remove unused methods and minimize cache reads in QueryManager's getCurrentResult [PR 2035](https://github.com/apollographql/apollo-client/pull/2035)
 - When cache implementations broadcast invalidations, they also provide the latest data for the invalidated query to minimize reads inside the QueryManager [PR 2031](https://github.com/apollographql/apollo-client/pull/2031)
 - Move abstract cache into its own module [PR #2019](https://github.com/apollographql/apollo-client/pull/2019)